chore: upgrade to cdk v2
Updating cdk to v2. We are currently publishing under 0.x.x, but have
published under major version 1, 2, and 3 in the past so this PR also
bumps the version to 4.

This PR also updates the tests to use assertions, but does keep the
snapshot test in order to include the snapshot diff in the PR.

BREAKING CHANGE: updates the dependency on aws-cdk to ^2.0.0-rc.28